pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "figgen"
version = "0.1.0"
description = "Publication-grade figure generation engine for offshore/geotechnical/structural engineering research"
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "KSK" }]

dependencies = [
    "matplotlib>=3.8",
    "numpy>=1.24",
    "pandas>=2.0",
    "pyarrow>=14.0",
    "scipy>=1.11",
    "openpyxl>=3.1",
    "pyyaml>=6.0",
    "jinja2>=3.1",
    "SciencePlots>=2.1",
    "cmocean>=3.0",
    "palettable>=3.3",
    "xarray>=2023.10",
    "pint>=0.23",
    "GitPython>=3.1",
    "Pillow>=10.0",
    "lxml>=4.9",
    "pyoxipng>=9.0",
]

[project.optional-dependencies]
shm = ["obspy>=1.4"]
mesh = ["pyvista>=0.42"]
latex = ["tikzplotlib>=0.10"]
agents = [
    "anthropic>=0.39",
]
svg = [
    "drawsvg>=2.4",
    "cairosvg>=2.7",
]
docs = [
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-glightbox>=0.4",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "pymdown-extensions>=10.7",
]
dev = [
    "pytest>=7.4",
    "pytest-mpl>=0.17",
    "pre-commit>=3.5",
    "ruff>=0.5",
    "mkdocs>=1.6",
    "mkdocs-material>=9.5",
    "mkdocs-glightbox>=0.4",
    "mkdocs-git-revision-date-localized-plugin>=1.2",
    "pymdown-extensions>=10.7",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-dir]
"" = "src"

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-ra --strict-markers"
filterwarnings = ["ignore::DeprecationWarning"]
markers = [
    "integration: end-to-end pipeline tests that spawn subprocesses (matplotlib backend, etc.)",
    "vision: tests that exercise the Anthropic vision API (require ANTHROPIC_API_KEY)",
]

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "N", "UP", "B"]
ignore = ["E501"]
